delete user bug fix

+++ Backend/UNUS/src/main/java/com/_an_5/UNUS/Users/UserController.java
@@ -62,7 +62,7 @@ public class UserController {
     @ApiOperation(value = "delete a user", response = String.class, tags = "user-controller")
     @DeleteMapping(path = "/user/{id}")
     public String deleteUser(@PathVariable int id) {
-        if(!userRepository.existsById((long)id))
+        if(!userRepository.existsById(id))
             return failure;
         userRepository.deleteById(id);
         return success;
